Add files via upload
This project involves creating bar charts to visualize the distribution of a categorical or continuous variable, such as age or gender, within a dataset. The visualization helps in understanding the frequency and distribution of different categories or numerical ranges. The chart was generated using tools/libraries like  Python, numpy, pandas, and  Matplotlib, . This task enhances data interpretation by providing a clear graphical representation of the data distribution.